old-tests/submission/Infraware/Forms/contents/Forms/Not Use/button_autofocus.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_autofocus</h3>
	</p>

	<li>The text control be focused when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>
	
	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><button id='button_id'>button</button></p>
	</form>

	<script>
	
		var button = document.getElementById("button_id");
		button.autofocus = true;

		if (typeof(button.autofocus) == "boolean") {
			if (button.autofocus != true) {
				test(function() {
					assert_true(false, "autofocus attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("autofocus attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
